requirements.txt
backcall==0.2.0
click==7.1.2
colorama==0.4.5
decorator==5.0.9
Flask-WTF==1.0.1
gunicorn==20.1.0
importlib-metadata==4.8.3
ipykernel==5.5.5
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
mariadb==1.0.9
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysql-connector-python==8.0.31
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
protobuf==3.19.6
ptyprocess==0.7.0
Pygments==2.9.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pywin32==304
pyzmq==24.0.1
six==1.16.0
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==3.0.0
zipp==3.6.0
pyjwt==2.6.0
bcrypt==4.0.1
pydantic
cryptography
sqlalchemy
python-dotenv